remove util.py
